Much jupyter fumbling to try and get kernel restarts working smoothly. They mostly do now but because the cache reference doesn't change a Ctrl-Alt-F9 can be needed to update Excel.
The code is still messy and would beneift from further work. I've included quite a few comments as the jupyter documentation is rather sparse and the canonical way to interact with the kernel is unclear.

tests: would need quite a advanced test harness to test the interaction of jupyter and Excel!