Fix: File source watch for changes.  This was broken by someone setting the watch subtree flag to false in ReadDirectoryChanges.  Perhaps counterinutively, this means the function only returns changes in the directory object itself, not its top level contents.
Watching the whole directory potentially routes a lot of APCs through Excel's main thread, so we found another library which uses the IO completion port variant of ReadDirectoryChanges - this happens in a background thread.  This has the further advantage that changes are detected immediately, without having to bring Excel in focus.