Do not wait for the COM interface before loading plugins.  This means xlOil loads earlier, eliminating delay when the first workbook is opened.  It also means all worksheet functions are registered before any workbook is open, avoiding #NAME errors as per GitHub #130.
The disadvantage is that any plugins or python code which attempt to use the COM interface during load will fail.  The obvious culprit is the Ribbon, so the *ExcelGUI* object has been changed to schedule its add-in connect rather than block in the constructor.  This probably covers most use cases.