Re-raise pulbished exception in the rtd subscribers
This is a change of behaviour so the package version should be incremented.

This makes it much easier and more pythonic to manage error states in calling code. Otherwise we have to check the value is a string which matches a pattern.

The end result in excel is that an error string is printed. However there is one more layer of wrapping which is occuring I believe.

Implement error raising using the latest ABI functions alongside the ones for older versions of python.